cast operand instead of result

+++ src/sq_idx.c
@@ -332,7 +332,7 @@ static int near _SquishAppendIndexRecord(HIDX hix, SQIDX * psqi)
     /* If we arrived here, we either have no segments, or all of our          *
      * existing segments are full.  To handle this, we need to reallocate     *
      * the array of pointers to segments and add a new one.                   */
-    pss = palloc(sizeof(SQIDXSEG) * (size_t)(hix->cSeg + 1));
+    pss = palloc(sizeof(SQIDXSEG) * ((size_t)(hix->cSeg) + 1));
 
     if(pss == NULL)
     {
@@ -601,7 +601,7 @@ int _SquishEndBuffer(HIDX hix)
 
                 if((long)dwStart + (long)hix->pss[i].dwUsed > hix->lDeltaHi)
                 {
-                    size = (size_t)(hix->lDeltaHi - (long)dwStart + 1L);
+                    size = (size_t)(hix->lDeltaHi) - (size_t)dwStart + 1L;
                 }
                 else
                 {
